script/requirements.txt
asn1crypto==0.24.0
astropy==3.1.2
atomicwrites==1.1.5
attrs==18.2.0
blist==1.3.6
BTrees==4.3.1
configobj==5.0.6
cryptography==2.6.1
Cython==0.29.2
decorator==4.3.0
entrypoints==0.3
gevent==1.3.7
greenlet==0.4.15
grpcio==1.16.1
gssapi==1.4.1
hidapi==0.7.99.post21
http-parser==0.8.3
indexed-gzip==0.8.6
ipython==5.8.0
ipython-genutils==0.2.0
Jinja2==2.10
jpy==0.9.0
JPype1==0.6.3
kdtree==0.7.1
keyring==17.1.1
keyrings.alt==3.1.1
Kyoto-Cabinet==1.5
MarkupSafe==1.1.0
more-itertools==4.2.0
numpy==1.16.2
olefile==0.46
opencv-python==4.5.1.48
persistent==4.2.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.8.0
ply==3.11
prelude==4.1.0
prompt-toolkit==1.0.15
psutil==5.5.1
py==1.7.0
py-radix==0.10.0
pyalsa==1.1.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
PyBluez==0.22
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.30.4
PyICU==2.2
pymad==0.10
pyodbc==4.0.22
pypcap==1.2.2
pyraf==2.1.15
pysendfile==2.0.1
pySFML==2.2.0
pysmbc==1.0.15.6
pytango==9.2.5
pytest==3.10.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.2.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-casacore==3.0.0
python-ldap==3.1.0
python-sane==2.8.3
pyxattr==0.6.1
pyxdg==0.25
qrencode==1.2
rpy2==2.9.5
scipy==1.1.0
SecretStorage==2.3.1
simplegeneric==0.8.1
six==1.12.0
stsci.tools==3.4.13
subvertpy==0.10.1
traitlets==4.3.2
urwid==2.0.1
wcwidth==0.1.7
wheezy.template==0.1.167
yenc==0.4.0
zodbpickle==1.0
zope.interface==4.3.2
